+++ packages/components/src/validated-form-controls/components/stories/input-control.story.tsx
@@ -22,7 +22,8 @@ import InputControlSuffixWrapper from '../../../input-control/input-suffix-wrapp
 import { Button } from '../../../button';
 
 const meta: Meta< typeof ValidatedInputControl > = {
-	title: 'Components (Experimental)/Validated Form Controls/ValidatedInputControl',
+	title: 'Components/Selection & Input/Validated Form Controls/ValidatedInputControl',
+	id: 'components-validatedinputcontrol',
 	component: ValidatedInputControl,
 	tags: [ 'status-private' ],
 	decorators: formDecorator,