+++ packages/components/src/validated-form-controls/components/stories/range-control.story.tsx
@@ -15,7 +15,8 @@ import { formDecorator } from './story-utils';
 import { ValidatedRangeControl } from '../range-control';
 
 const meta: Meta< typeof ValidatedRangeControl > = {
-	title: 'Components (Experimental)/Validated Form Controls/ValidatedRangeControl',
+	title: 'Components/Selection & Input/Validated Form Controls/ValidatedRangeControl',
+	id: 'components-validatedrangecontrol',
 	component: ValidatedRangeControl,
 	tags: [ 'status-private' ],
 	decorators: formDecorator,